Refactored CurrencyHelper.cs
