--- FinanceTrackerCore/TestData/apiResponse.json
@@ -1,178 +1,0 @@
-{
-  "success": true,
-  "timestamp": 1750768744,
-  "base": "EUR",
-  "date": "2025-06-24",
-  "rates": {
-    "AED": 4.26307,
-    "AFN": 82.426603,
-    "ALL": 98.184226,
-    "AMD": 447.265546,
-    "ANG": 2.077429,
-    "AOA": 1064.471906,
-    "ARS": 1359.90611,
-    "AUD": 1.784668,
-    "AWG": 2.092377,
-    "AZN": 1.975915,
-    "BAM": 1.956448,
-    "BBD": 2.342347,
-    "BDT": 141.877013,
-    "BGN": 1.956912,
-    "BHD": 0.437841,
-    "BIF": 3455.078828,
-    "BMD": 1.16082,
-    "BND": 1.486037,
-    "BOB": 8.033801,
-    "BRL": 6.373014,
-    "BSD": 1.160119,
-    "BTC": 0.000011046588,
-    "BTN": 99.770068,
-    "BWP": 15.457388,
-    "BYN": 3.796573,
-    "BYR": 22752.065217,
-    "BZD": 2.330262,
-    "CAD": 1.592256,
-    "CDF": 3339.677996,
-    "CHF": 0.938389,
-    "CLF": 0.028585,
-    "CLP": 1096.846848,
-    "CNY": 8.334104,
-    "CNH": 8.323228,
-    "COP": 4740.787467,
-    "CRC": 587.212286,
-    "CUC": 1.16082,
-    "CUP": 30.761721,
-    "CVE": 110.30155,
-    "CZK": 24.785067,
-    "DJF": 206.586464,
-    "DKK": 7.459474,
-    "DOP": 68.652157,
-    "DZD": 150.664627,
-    "EGP": 58.095773,
-    "ERN": 17.412295,
-    "ETB": 159.479428,
-    "EUR": 1,
-    "FJD": 2.609232,
-    "FKP": 0.860571,
-    "GBP": 0.853322,
-    "GEL": 3.15738,
-    "GGP": 0.860571,
-    "GHS": 12.007186,
-    "GIP": 0.860571,
-    "GMD": 82.99648,
-    "GNF": 10048.329656,
-    "GTQ": 8.920956,
-    "GYD": 242.709834,
-    "HKD": 9.11226,
-    "HNL": 30.30178,
-    "HRK": 7.535224,
-    "HTG": 152.259142,
-    "HUF": 401.891987,
-    "IDR": 18964.426768,
-    "ILS": 3.947803,
-    "IMP": 0.860571,
-    "INR": 99.895503,
-    "IQD": 1519.690479,
-    "IRR": 48899.52777,
-    "ISK": 142.212086,
-    "JEP": 0.860571,
-    "JMD": 185.501469,
-    "JOD": 0.823023,
-    "JPY": 168.30145,
-    "KES": 149.943296,
-    "KGS": 101.3009,
-    "KHR": 4650.581098,
-    "KMF": 496.253612,
-    "KPW": 1044.727432,
-    "KRW": 1578.679811,
-    "KWD": 0.354975,
-    "KYD": 0.966729,
-    "KZT": 601.237356,
-    "LAK": 25027.263336,
-    "LBP": 103949.936251,
-    "LKR": 348.541499,
-    "LRD": 232.015876,
-    "LSL": 20.565415,
-    "LTL": 3.427598,
-    "LVL": 0.702168,
-    "LYD": 6.285083,
-    "MAD": 10.559181,
-    "MDL": 19.790728,
-    "MGA": 5157.909096,
-    "MKD": 61.557519,
-    "MMK": 2437.109873,
-    "MNT": 4158.471221,
-    "MOP": 9.380272,
-    "MRU": 45.833187,
-    "MUR": 52.678385,
-    "MVR": 17.88246,
-    "MWK": 2011.64457,
-    "MXN": 22.099162,
-    "MYR": 4.926507,
-    "MZN": 74.245725,
-    "NAD": 20.565237,
-    "NGN": 1798.608636,
-    "NIO": 42.693963,
-    "NOK": 11.723298,
-    "NPR": 159.631709,
-    "NZD": 1.926305,
-    "OMR": 0.446309,
-    "PAB": 1.160129,
-    "PEN": 4.166381,
-    "PGK": 4.781667,
-    "PHP": 66.102294,
-    "PKR": 330.564633,
-    "PLN": 4.249572,
-    "PYG": 9261.028882,
-    "QAR": 4.229301,
-    "RON": 5.048171,
-    "RSD": 117.219557,
-    "RUB": 91.130206,
-    "RWF": 1675.169522,
-    "SAR": 4.354605,
-    "SBD": 9.681733,
-    "SCR": 16.404697,
-    "SDG": 697.075344,
-    "SEK": 11.071375,
-    "SGD": 1.485576,
-    "SHP": 0.912222,
-    "SLE": 26.060034,
-    "SLL": 24341.811987,
-    "SOS": 663.039699,
-    "SRD": 45.087363,
-    "STD": 24026.623151,
-    "SVC": 10.150363,
-    "SYP": 15092.896237,
-    "SZL": 20.548606,
-    "THB": 37.962867,
-    "TJS": 11.432832,
-    "TMT": 4.062869,
-    "TND": 3.419833,
-    "TOP": 2.718762,
-    "TRY": 46.011991,
-    "TTD": 7.88385,
-    "TWD": 34.24886,
-    "TZS": 3102.297033,
-    "UAH": 48.43565,
-    "UGX": 4174.509148,
-    "USD": 1.16082,
-    "UYU": 47.055593,
-    "UZS": 14397.770904,
-    "VES": 120.268834,
-    "VND": 30378.650344,
-    "VUV": 139.464419,
-    "WST": 3.085352,
-    "XAF": 656.180087,
-    "XAG": 0.032507,
-    "XAU": 0.000351,
-    "XCD": 3.137173,
-    "XDR": 0.816457,
-    "XOF": 656.182915,
-    "XPF": 119.331742,
-    "YER": 281.673178,
-    "ZAR": 20.577072,
-    "ZMK": 10448.77733,
-    "ZMW": 27.157467,
-    "ZWL": 373.783455
-  }
-}
